pyproject.toml
[project]
name = "extraction-engine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=1.2.15",
    "langchain-groq>=1.1.2",
    "langchain-mistralai>=1.1.2",
    "langchain-text-splitters>=1.1.1",
    "pandas>=3.0.2",
    "pathlib>=1.0.1",
    "python-dotenv>=1.2.2",
]
